+++ packages/sdks/angular-sdk/projects/demo-app/src/app/my-applications-portal/my-applications-portal.component.ts
@@ -1,0 +1,15 @@
+import { Component } from '@angular/core';
+import { environment } from '../../environments/environment';
+import { Router } from '@angular/router';
+
+@Component({
+  selector: 'app-my-applications-portal',
+  templateUrl: './my-applications-portal.component.html',
+  styleUrls: ['./my-applications-portal.scss']
+})
+export class MyApplicationsPortalComponent {
+  theme = (environment.descopeTheme as 'light' | 'dark' | 'os') ?? 'os';
+  debugMode = environment.descopeDebugMode ?? false;
+
+  constructor(private router: Router) {}
+}