+++ packages/sdks/web-component/src/lib/descope-wc/DescopeWc.ts
@@ -30,6 +30,7 @@ import {
   leadingDebounce,
   handleReportValidityOnBlur,
   getUserLocale,
+  clearPreviousExternalInputs,
 } from '../helpers';
 import { calculateConditions, calculateCondition } from '../helpers/conditions';
 import { getLastAuth, setLastAuth } from '../helpers/lastAuth';
@@ -186,6 +187,27 @@ class DescopeWc extends BaseDescopeWc {
     return null;
   }
 
+  async #handleFlowRestart() {
+    this.loggerWrapper.debug('Trying to restart the flow');
+    const prevCompVersion = await this.getComponentsVersion();
+    this.getConfig.reset();
+    const compVersion = await this.getComponentsVersion();
+
+    if (prevCompVersion === compVersion) {
+      this.loggerWrapper.debug(
+        'Components version was not changed, restarting flow',
+      );
+      this.flowState.update({
+        stepId: null,
+        executionId: null,
+      });
+    } else {
+      this.loggerWrapper.error(
+        'Components version mismatch, please reload the page',
+      );
+    }
+  }
+
   async onFlowChange(
     currentState: FlowState,
     prevState: FlowState,
@@ -225,14 +247,20 @@ class DescopeWc extends BaseDescopeWc {
     const loginId = this.sdk.getLastUserLoginId();
     const flowConfig = await this.getFlowConfig();
     const projectConfig = await this.getProjectConfig();
-
+    const flowVersions = Object.entries(projectConfig.flows || {}).reduce( // pass also current versions for all flows, it may be used as a part of the current flow
+      (acc, [key, value]) => {
+        acc[key] = value.version;
+        return acc;
+      },
+      {} as Record<string, number>,
+    );
     const redirectAuth =
       redirectAuthCallbackUrl && redirectAuthCodeChallenge
         ? {
-            callbackUrl: redirectAuthCallbackUrl,
-            codeChallenge: redirectAuthCodeChallenge,
-            backupCallbackUri: redirectAuthBackupCallbackUri,
-          }
+          callbackUrl: redirectAuthCallbackUrl,
+          codeChallenge: redirectAuthCodeChallenge,
+          backupCallbackUri: redirectAuthBackupCallbackUri,
+        }
         : undefined;
 
     // if there is no execution id we should start a new flow
@@ -279,8 +307,8 @@ class DescopeWc extends BaseDescopeWc {
           },
           conditionInteractionId,
           '',
-          flowConfig.version,
           projectConfig.componentsVersion,
+          flowVersions,
           {
             ...this.formConfigValues,
             ...(code ? { exchangeCode: code, idpInitiated: true } : {}),
@@ -505,8 +533,8 @@ class DescopeWc extends BaseDescopeWc {
           },
           conditionInteractionId,
           interactionId,
-          version,
           componentsVersion,
+          flowVersions,
           {
             ...this.formConfigValues,
             ...inputs,
@@ -610,6 +638,16 @@ class DescopeWc extends BaseDescopeWc {
         sdkResp?.error?.errorDescription || defaultMessage,
         sdkResp?.error?.errorMessage || defaultDescription,
       );
+
+      // E102004 = Flow requested is in old version
+      // E103205 = Flow timed out
+      const errorCode = sdkResp?.error?.errorCode
+      if (
+        (errorCode === 'E102004' || errorCode === 'E103205') &&
+        this.isRestartOnError
+      ) {
+        this.#handleFlowRestart();
+      }
       return;
     }
 
@@ -842,19 +880,23 @@ class DescopeWc extends BaseDescopeWc {
 
       this.rootElement.replaceChildren(clone);
 
-      // we need to wait for all components to render before we can set its value
-      setTimeout(() => {
-        updateScreenFromScreenState(this.rootElement, screenState);
-      });
-
       // If before html url was empty, we deduce its the first time a screen is shown
       const isFirstScreen = !prevState.htmlUrl;
 
-      handleAutoFocus(this.rootElement, this.autoFocus, isFirstScreen);
+      // we need to wait for all components to render before we can set its value
+      setTimeout(() => {
+        updateScreenFromScreenState(this.rootElement, screenState);
+        this.#updateExternalInputs();
 
-      if (this.validateOnBlur) {
-        handleReportValidityOnBlur(this.rootElement);
-      }
+        handleAutoFocus(this.rootElement, this.autoFocus, isFirstScreen);
+
+        if (this.validateOnBlur) {
+          handleReportValidityOnBlur(this.rootElement);
+        }
+
+        // we need to wait for all components to render before we can set its value
+        updateScreenFromScreenState(this.rootElement, screenState);
+      });
 
       this.#hydrate(next);
       if (isFirstScreen) {
@@ -972,6 +1014,37 @@ class DescopeWc extends BaseDescopeWc {
         this.loggerWrapper.error('Could not store credentials', e.message);
       }
     }
+  }
+
+  #updateExternalInputs() {
+    // we need to clear external inputs that were created previously, so each screen has only
+    // the slotted inputs it needs
+    clearPreviousExternalInputs();
+
+    const eles = this.rootElement.querySelectorAll('[external-input="true"]');
+    eles.forEach((ele) => this.#handleExternalInputs(ele));
+  }
+
+  #handleExternalInputs(ele: Element) {
+    if (!ele) {
+      return;
+    }
+
+    const origInputs = ele.querySelectorAll('input');
+
+    origInputs.forEach((inp) => {
+      const targetSlot = inp.getAttribute('slot');
+      const id = `input-${ele.id}-${targetSlot}`;
+
+      const slot = document.createElement('slot');
+      slot.setAttribute('name', id);
+      slot.setAttribute('slot', targetSlot);
+
+      ele.appendChild(slot);
+
+      inp.setAttribute('slot', id);
+      this.appendChild(inp);
+    });
   }
 
   // we are wrapping this function with a leading debounce,