+++ packages/sdks/web-component/src/lib/helpers/helpers.ts
@@ -366,14 +366,22 @@ const compareArrays = (array1: any[], array2: any[]) =>
 export const withMemCache = <I extends any[], O>(fn: (...args: I) => O) => {
   let prevArgs: any[];
   let cache: any;
-  return (...args: I) => {
-    if (prevArgs && compareArrays(prevArgs, args)) return cache as O;
+  return Object.assign(
+    (...args: I) => {
+      if (prevArgs && compareArrays(prevArgs, args)) return cache as O;
 
-    prevArgs = args;
-    cache = fn(...args);
+      prevArgs = args;
+      cache = fn(...args);
 
-    return cache as O;
-  };
+      return cache as O;
+    },
+    {
+      reset: () => {
+        prevArgs = undefined;
+        cache = undefined;
+      },
+    },
+  );
 };
 
 export const handleAutoFocus = (
@@ -549,3 +557,9 @@ export function getUserLocale(locale: string): Locale {
 
   return { locale: nl.toLowerCase(), fallback: nl.toLowerCase() };
 }
+
+export const clearPreviousExternalInputs = () => {
+  document
+    .querySelectorAll('[data-hidden-input="true"]')
+    .forEach((ele) => ele.remove());
+};