+++ packages/widgets/audit-management-widget/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.89](https://github.com/descope/descope-js/compare/audit-management-widget-0.1.88...audit-management-widget-0.1.89) (2024-10-14)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.18.0`
+## [0.1.88](https://github.com/descope/descope-js/compare/audit-management-widget-0.1.87...audit-management-widget-0.1.88) (2024-09-29)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.17.0`
 ## [0.1.87](https://github.com/descope/descope-js/compare/audit-management-widget-0.1.86...audit-management-widget-0.1.87) (2024-09-19)
 
 ### Dependency Updates