+++ packages/widgets/role-management-widget/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.123](https://github.com/descope/descope-js/compare/role-management-widget-0.1.122...role-management-widget-0.1.123) (2024-10-14)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.18.0`
+## [0.1.122](https://github.com/descope/descope-js/compare/role-management-widget-0.1.121...role-management-widget-0.1.122) (2024-09-29)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.17.0`
 ## [0.1.121](https://github.com/descope/descope-js/compare/role-management-widget-0.1.120...role-management-widget-0.1.121) (2024-09-19)
 
 ### Dependency Updates